requirements.txt
asgiref==3.7.2
certifi==2024.2.2
Django==4.2.0
django-debug-toolbar==4.3.0
django-filter==24.2
django-request-logging==0.7.5
djangorestframework==3.15.1
drf-spectacular==0.27.2
httpx==0.27.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
python-dotenv==1.0.1
PyYAML==6.0.1
sentry-sdk==1.45.0
sitemaps==0.1.0
sqlparse==0.4.4
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
